Deleted entry point and created main.cpp in application project for simplicity. Turned Game object into a unique pointer instead of declaring directly on the heap so we dont have memory leaks in case of an exception. Changed the main window position for convenience. General cleanup of code.
